Convert app.js to typescript
